+++ mrgeo-core/src/main/java/org/mrgeo/hdfs/vector/shp/esri/geom/JPolygon.java
@@ -302,7 +302,7 @@ public double getArea()
     area = tempArea;
   }
 
-  return area.doubleValue();
+  return area;
 }
 
 public Coord getCentroid()
