+++ mrgeo-core/src/main/java/org/mrgeo/hdfs/vector/shp/esri/geom/JPolygonZ.java
@@ -289,7 +289,7 @@ public double getArea()
     area = tempArea;
   }
 
-  return area.doubleValue();
+  return area;
 }
 
 public Coord getCentroid()
